+++ helpers/NPMoveItems.js
@@ -259,14 +259,13 @@ export function moveParagraphToNote(para: TParagraph, destinationNote: TNote): b
  * Move a given paragraph (and any following indented paragraphs) to a different note.
  * Note: simplified version of 'moveParas()' in NPParagraph.
  * NB: the Setting 'includeFromStartOfSection' decides whether these directly following paragaphs have to be indented (false) or can take all following lines at same level until next empty line as well.
- * Note: originally in helpers/blocks.js, not used anywhere yet.
  * @param {TParagraph} para
  * @param {string} destFilename
  * @param {NoteType} destNoteType
  * @param {string} destHeading to move under
  * @author @jgclark
  */
-export function moveGivenParaAndBlock(para: TParagraph, destFilename: string, destNoteType: NoteType, destHeading: string): void {
+export function moveGivenParaAndIndentedChildren(para: TParagraph, destFilename: string, destNoteType: NoteType, destHeading: string): void {
   try {
     if (!destFilename) {
       throw new Error('Invalid destination filename given.')
@@ -282,7 +281,7 @@ export function moveGivenParaAndBlock(para: TParagraph, destFilename: string, de
 
     // get children paras (as well as the original)
     const parasInBlock = getParaAndAllChildren(para)
-    logDebug('blocks/moveGivenParaAndBlock', `moveParas: move block of ${parasInBlock.length} paras`)
+    logDebug('blocks/moveGivenParaAndIndentedChildren', `moveParas: move block of ${parasInBlock.length} paras`)
 
     // Note: There's still no API function to add multiple
     // paragraphs in one go, but we can insert a raw text string.
@@ -293,14 +292,14 @@ export function moveGivenParaAndBlock(para: TParagraph, destFilename: string, de
     if (!destNote) {
       throw new Error(`Destination note can't be found from filename '${destFilename}'`)
     }
-    logDebug('blocks/moveGivenParaAndBlock', `- Moving to note '${displayTitle(destNote)}' under heading: '${destHeading}'`)
+    logDebug('blocks/moveGivenParaAndIndentedChildren', `- Moving to note '${displayTitle(destNote)}' under heading: '${destHeading}'`)
     addParasAsText(destNote, selectedParasAsText, destHeading, 'start', true)
 
     // delete from existing location
-    logDebug('blocks/moveGivenParaAndBlock', `- Removing ${parasInBlock.length} paras from original note`)
+    logDebug('blocks/moveGivenParaAndIndentedChildren', `- Removing ${parasInBlock.length} paras from original note`)
     originNote.removeParagraphs(parasInBlock)
   }
   catch (error) {
-    logError('blocks/moveGivenParaAndBlock', `moveParas(): ${error.message}`)
+    logError('blocks/moveGivenParaAndIndentedChildren', `moveParas(): ${error.message}`)
   }
 }